Further restructuring, remove unused imports
