+++ Source/Event.swift
@@ -7,7 +7,6 @@
 // This software may be modified and distributed under the terms of the MIT license.
 // See the LICENSE file for details.
 //
-import Foundation
 
 /**
  The events that can be dispatched by the task.
